+++ src/Apps/Order/Components/ShippingArtaSummaryItem.tsx
@@ -7,7 +7,7 @@ import * as React from "react"
 import { createFragmentContainer, graphql } from "react-relay"
 import { Text } from "@artsy/palette"
 import { startCase } from "lodash"
-import { shippingQuoteDisplayNames } from "Apps/Order/Routes/Shipping/Components/ShippingQuotes2"
+import { shippingQuoteDisplayNames } from "Apps/Order/Routes/Shipping/Components/ShippingQuotes"
 
 interface ShippingArtaSummaryItemProps {
   order: ShippingArtaSummaryItem_order$data
